+++ python/ascii_stream_engine/adapters/processors/filters/__init__.py
@@ -1,23 +1,39 @@
 from .base import BaseFilter
+from .boids import BoidsFilter
 from .brightness import BrightnessFilter
 from .cpp_brightness_contrast import CppBrightnessContrastFilter
 from .cpp_channel_swap import CppChannelSwapFilter
 from .cpp_grayscale import CppGrayscaleFilter
 from .cpp_invert import CppInvertFilter
 from .cpp_modifier import CppImageModifierFilter
+from .cpp_physarum import CppPhysarumFilter
 from .detail import DetailBoostFilter
+from .edge_smooth import EdgeSmoothFilter
 from .edges import EdgeFilter
 from .invert import InvertFilter
+from .optical_flow_particles import OpticalFlowParticlesFilter
+from .physarum import PhysarumFilter
+from .radial_collapse import RadialCollapseFilter
+from .stippling import StipplingFilter
+from .uv_displacement import UVDisplacementFilter
 
 __all__ = [
     "BaseFilter",
+    "BoidsFilter",
     "BrightnessFilter",
     "CppBrightnessContrastFilter",
     "CppChannelSwapFilter",
     "CppGrayscaleFilter",
     "CppImageModifierFilter",
     "CppInvertFilter",
+    "CppPhysarumFilter",
     "DetailBoostFilter",
     "EdgeFilter",
+    "EdgeSmoothFilter",
     "InvertFilter",
+    "OpticalFlowParticlesFilter",
+    "PhysarumFilter",
+    "RadialCollapseFilter",
+    "StipplingFilter",
+    "UVDisplacementFilter",
 ]